+++ src/components/Utility.tsx
@@ -1,30 +1,11 @@
-import { useState } from 'react';
-import { TableBody, TableCell, TableRow } from '@mui/material';
-import { TableComponent } from '@/components/Table/TableComponent';
-import { useGetAllUtilityDataQuery } from '@/state/genability/genabilitySlice';
-import { utilityRows } from '@/utils';
+import { useGetSingleUtilityBasedOnLseIdQuery } from '@/state/genability/genabilitySlice';
 
-export const Utility = () => {
-  const [page, setPage] = useState(0);
-  const [rowsPerPage, setRowsPerPage] = useState(10);
-  const [fetchPage, setFetchPage] = useState(0);
-  const { data, isLoading, isError } = useGetAllUtilityDataQuery({ pageStart: fetchPage, pageCount: rowsPerPage });
+type UtilityProps = {
+  lseId: string;
+};
 
-  const handleChangePage = (_e: React.MouseEvent<HTMLButtonElement> | null, newPage: number) => {
-    if (!data) return;
-
-    const nextPage = newPage * rowsPerPage;
-    setPage(newPage);
-    setFetchPage(nextPage);
-  };
-
-  const handleChangeRowsPerPage = (event: React.ChangeEvent<HTMLInputElement>) => {
-    if (!data) return;
-
-    setRowsPerPage(+event.target.value);
-    setFetchPage(0);
-    setPage(0);
-  };
+export const Utility = ({ lseId }: UtilityProps) => {
+  const { data, isLoading, isError } = useGetSingleUtilityBasedOnLseIdQuery(lseId);
 
   if (isLoading) {
     return <div>Loading...</div>;
@@ -35,31 +16,19 @@ export const Utility = () => {
   }
 
   return (
-    <TableComponent
-      headerRows={utilityRows}
-      data={data}
-      page={page}
-      handleChangePage={handleChangePage}
-      handleChangeRowsPerPage={handleChangeRowsPerPage}
-      title="List of Utilities">
-      <TableBody>
-        {data.results.map((lse) => (
-          <TableRow
-            data-testid={`cy-row-${lse.lseId}`}
-            hover
-            role="checkbox"
-            tabIndex={-1}
-            key={lse.lseId}
-            sx={{ cursor: 'pointer' }}>
-            <TableCell className="min-w-xs">{lse.name}</TableCell>
-            <TableCell>{lse.lseCode}</TableCell>
-            <TableCell className="max-w-xs break-words">{lse.websiteHome}</TableCell>
-            <TableCell align="center">
-              {lse.totalCustomers !== null ? lse.totalCustomers.toLocaleString('en-US') : null}
-            </TableCell>
-          </TableRow>
-        ))}
-      </TableBody>
-    </TableComponent>
+    <section className="flex-1 h-screen overflow-auto">
+      <div className="bg-[#f5f5f5] py-4 px-6">
+        <h1 className="text-2xl">
+          {data?.results[0].name} ({data?.results[0].lseCode})
+        </h1>
+        <a
+          className="font-thin text-sky-700 hover:underline"
+          href={data?.results[0].websiteHome}
+          target="_blank"
+          rel="noreferrer">
+          {data?.results[0].websiteHome}
+        </a>
+      </div>
+    </section>
   );
 };
