+++ src/pages/LoadEntity.tsx
@@ -1,9 +1,9 @@
-import { Utility } from '@/components/Utility';
+import { Utilities } from '@/components/Utilities';
 
 export const LoadEntity = () => {
   return (
     <div className="flex-[0.8_1_0%] h-screen">
-      <Utility />
+      <Utilities />
     </div>
   );
 };
